[PP][OPS-9970] moved model and util classes to the cor
