+++ app/model/Journey.scala
@@ -16,10 +16,11 @@
 
 package model
 
-import model.pcipal.{PcipalSessionId, PcipalSessionLaunchRequest, PcipalSessionLaunchResponse}
+import model.pcipal.{PcipalSessionId, PcipalSessionLaunchResponse}
 import play.api.libs.json.{Json, OFormat}
 import tps.model.{JourneyId, Navigation}
 import tps.model.repo.HasId
+import tps.pcipalmodel.PcipalSessionLaunchRequest
 
 import java.time.Instant
