+++ app/model/TpsPaymentItem.scala
@@ -18,7 +18,7 @@ package model
 
 import model.pcipal.ChargeRefNotificationPcipalRequest
 import play.api.libs.json._
-import tps.model.HeadOfDutyIndicator
+import tps.model.{HeadOfDutyIndicator, PaymentItemId}
 
 import java.time.Instant
