+++ app/paymentsprocessor/PaymentsProcessorService.scala
@@ -16,8 +16,9 @@
 
 package paymentsprocessor
 
-import model.{MibSpecificData, PaymentItemId}
+import model.MibSpecificData
 import repository.JourneyRepo
+import tps.model.PaymentItemId
 
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.{ExecutionContext, Future}
